Update README to include notes for setting emulator env var (#10)
# Existing Sample Changes
## Description

Previously the emulator host pipe depth were defaulted to infinite, which doesn't match correct bahvior in documentation. I fixed the emulator pipe depth control environment variable `CL_CONFIG_CHANNEL_DEPTH_EMULATION_MODE` to default to `strict` mode (it will default to depth 1 if not specified), so if a design wishes to increase the host buffer capacity (in emulation) of a host pipe to larger than the buffer on the device, `CL_CONFIG_CHANNEL_DEPTH_EMULATION_MODE` will have to be set to `ignore-depth`.

This change updates the README's of designs that do exactly that.

Fixes Issue# 

## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?

Please describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. Please also list any relevant details for your test configuration

**_Delete this line and everything below if this is not a PR for a new code sample_**

**_Delete this line and all above it if this PR is for a new code sample_**
# Adding a New Sample(s)

## Description

Please include a description of the sample

## Checklist
Administrative
- [ ] Review sample design with the repository owners: yuguen-intel, jessicadavies-intel

Code Development
- [ ] Implement coding guidelines and ensure code quality. [see wiki for details](https://github.com/altera-fpga/hls-samples/wiki/General-Code-Guidelines)
- [ ] Adhere to readme template 
- [ ] Enforce format via clang-format
- [ ] Adhere to sample.json specification. https://github.com/altera-fpga/hls-samples/wiki/sample-json-specification
- [ ] Run jsonlint on sample.json to verify json syntax. www.jsonlint.com

# Linked PRs:
- altera-innersource/applications.fpga.oneapi.products.acl-source#2637
- altera-innersource/applications.fpga.oneapi.products.acl-test#2930
- altera-innersource/applications.fpga.oneapi.products.hls-samples-source#10
- altera-innersource/applications.fpga.oneapi.stella#13389