Replace `-no-finite-math-only` flag with `-honor-infs` + `-honor-nans` flags case:30020113132 (#13)
## Description

 **Replace -no-finite-math-only with -honor-infs + honor-nans in `cholesky_inversion`,  `qrd`, `qri`, and `matmul ` code samples**.


**Fixes HSDES case:30020113132.**

The [commit c9c91f5](https://github.com/intel/llvm/commit/c9c91f59c318e28d884e5762f303348ba5724d21) from intel/llvm removes the `-no-finite-math-only` flag, and based on the documentation one should instead use the conjuction of `-honor-infs` and `-honor-nans`. 

Some of the code sample `CMakeList.txt` files included `-no-finite-math-only` (hence the warning and in turn the error) and so this pr makes the change described above.

## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?

The affected tests are now passing, as seen with the following SPETC runs:

- matmul: https://spetc.altera.com/testsummary?testRunIds=11776584
- qrd: https://spetc.altera.com/testsummary?testRunIds=11776587
- qri: https://spetc.altera.com/testsummary?testRunIds=11776588
- cholesky_inversion/complex_32_32: https://spetc.altera.com/testsummary?testRunIds=11776591
- cholesky_inversion/unmodified: https://spetc.altera.com/testsummary?testRunIds=11776595

# Linked PRs:
- altera-innersource/applications.fpga.oneapi.products.acl-test#2945
- altera-innersource/applications.fpga.oneapi.products.hls-samples-source#13
- altera-innersource/applications.fpga.oneapi.stella#13550