Fix  `-no-finite-math-only`  problem (ctd) (#21)
## Description

This is an extension of [this PR](https://github.com/altera-innersource/applications.fpga.oneapi.products.hls-samples-source/pull/13) which replaces `-no-finite-math-only` with `-honor-infs` + `-honor-nans` in in the CMakeLists.txt files of some code samples. 

It turns out there were two more samples that required this same change: `pca` and `cholesky`.

## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?

For cholesky, I followed the instructions on `products/source/hls-samples/ReferenceDesigns/cholesky/sample.json` for Linux and the test passes.

For pca, I followed the instructions on `products/source/hls-samples/ReferenceDesigns/pca/sample.json` for Linux and again all tests pass.

# Linked PRs:
- altera-innersource/applications.fpga.oneapi.products.hls-samples-source#21
- altera-innersource/applications.fpga.oneapi.stella#13616