remove sample.jsons (#32)
# Existing Sample Changes
## Description
cherry-pick #28, plus some additional samples that were removed in the eap branch
Remove sample.json for all samples, since they were only necessary for the oneAPI samples browser that we no longer support
## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?
regtests

# Linked PRs:
- altera-innersource/applications.fpga.oneapi.products.acl-test#3015
- altera-innersource/applications.fpga.oneapi.products.hls-samples-source#32
- altera-innersource/applications.fpga.oneapi.stella#13725