Update code samples to use uses_ready property (#53)
# Existing Sample Changes
## Description

We are separating the uses_ready host pipe property from the protocol property as this aligns with the pipe properties documentation. So also updating the code samples to reflect that.

## Type of change

Please delete options that are not relevant. Add a 'X' to the one that is applicable. 

- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)

## How Has This Been Tested?

See main acl-source PR for testing.

# Linked PRs:
- altera-innersource/applications.fpga.oneapi.products.acl-source#2729
- altera-innersource/applications.fpga.oneapi.products.acl-test#3052
- altera-innersource/applications.fpga.oneapi.products.hls-samples-source#53
- altera-innersource/applications.fpga.oneapi.stella#13845